+++ flutter_readium/lib/src/flutter_readium_web.dart
@@ -241,6 +241,11 @@ class FlutterReadiumWebPlugin extends FlutterReadiumPlatform {
   }
 
   @override
+  Future<void> setPDFPreferences(PDFPreferences preferences) async {
+    R2Log.d('setPDFPreferences is not supported on web platform');
+  }
+
+  @override
   Future<void> applyDecorations(String id, List<ReaderDecoration> decorations) async {
     R2Log.d('applyDecorations is not implemented on web platform');
   }